Refresh timeline when resetting statistics
