+++ event/src/main/java/de/danoeh/antennapod/event/StatisticsEvent.java
@@ -1,0 +1,7 @@
+package de.danoeh.antennapod.event;
+
+public class StatisticsEvent {
+
+    public StatisticsEvent() {
+    }
+}
