+++ ui/statistics/build.gradle
@@ -13,6 +13,7 @@ android {
 
 dependencies {
     implementation project(":core")
+    implementation project(':event')
     implementation project(":model")
     implementation project(":ui:common")
 
@@ -24,6 +25,7 @@ dependencies {
     implementation "androidx.viewpager2:viewpager2:$viewPager2Version"
     implementation "com.google.android.material:material:$googleMaterialVersion"
 
+    implementation "org.greenrobot:eventbus:$eventbusVersion"
     implementation "com.github.bumptech.glide:glide:$glideVersion"
     annotationProcessor "com.github.bumptech.glide:compiler:$glideVersion"
     implementation "io.reactivex.rxjava2:rxandroid:$rxAndroidVersion"
